requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pdf2image==1.12.1
pdfminer.six==20200104
pdfplumber==0.5.16
Pillow==7.0.0
pycryptodome==3.9.7
pytesseract==0.3.2
requests==2.23.0
six==1.14.0
sortedcontainers==2.1.0
unicodecsv==0.14.1
urllib3==1.25.8
Wand==0.5.9
